requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
pandas==2.1.0
scikit-learn==1.3.0
openai==0.28.0
numpy==1.25.2
python-dotenv==1.0.0
jinja2==3.1.2
